Adds frontend boilerplate and incident index
A set of files for defining a basic frontend structure were addded. Material-UI is used on favor of speed and cleanliness for the visual aspect of the project.

As for the project, an incident list is implemented fetching its contents from the API and displaying them into a table and also a map thanks to Leaftlet